Add more handling for triggers

+++ Maple2.Server.Game/Trigger/TriggerContext.Arcade.cs
@@ -4,13 +4,13 @@ public partial class TriggerContext {
     #region BoomBoomOcean
     public void ArcadeBoomBoomOceanSetSkillScore(int skillId, int score) { }
 
-    public void ArcadeBoomBoomOceanStartGame(byte lifeCount) { }
+    public void ArcadeBoomBoomOceanStartGame(int lifeCount) { }
 
     public void ArcadeBoomBoomOceanEndGame() { }
 
-    public void ArcadeBoomBoomOceanStartRound(byte round, int roundDuration, int timeScoreRate) { }
+    public void ArcadeBoomBoomOceanStartRound(int round, int roundDuration, int timeScoreRate) { }
 
-    public void ArcadeBoomBoomOceanClearRound(byte round) { }
+    public void ArcadeBoomBoomOceanClearRound(int round) { }
     #endregion
 
     #region SpringFarm
@@ -18,26 +18,26 @@ public partial class TriggerContext {
 
     public void ArcadeSpringFarmSpawnMonster(int[] spawnIds, int score) { }
 
-    public void ArcadeSpringFarmStartGame(byte lifeCount) { }
+    public void ArcadeSpringFarmStartGame(int lifeCount) { }
 
     public void ArcadeSpringFarmEndGame() { }
 
-    public void ArcadeSpringFarmStartRound(byte round, int uiDuration, string timeScoreType, int timeScoreRate, int roundDuration) { }
+    public void ArcadeSpringFarmStartRound(int round, int uiDuration, string timeScoreType, int timeScoreRate, int roundDuration) { }
 
-    public void ArcadeSpringFarmClearRound(byte round) { }
+    public void ArcadeSpringFarmClearRound(int round) { }
     #endregion
 
     #region ThreeTwoOne
-    public void ArcadeThreeTwoOneStartGame(byte lifeCount, int initScore) { }
+    public void ArcadeThreeTwoOneStartGame(int lifeCount, int initScore) { }
 
     public void ArcadeThreeTwoOneEndGame() { }
 
-    public void ArcadeThreeTwoOneStartRound(byte round, int uiDuration) { }
+    public void ArcadeThreeTwoOneStartRound(int round, int uiDuration) { }
 
     public void ArcadeThreeTwoOneResultRound(byte resultDirection) { }
 
-    public void ArcadeThreeTwoOneResultRound2(byte round) { }
+    public void ArcadeThreeTwoOneResultRound2(int round) { }
 
-    public void ArcadeThreeTwoOneClearRound(byte round) { }
+    public void ArcadeThreeTwoOneClearRound(int round) { }
     #endregion
 }
